Use self instead of README for stat check
Currently the script relies on README.md to be in the working directory, otherwise the script fails and incorrectly reports a language error. This is incompatible with the readme which says you can just download the script alone. Instead, this change applies the stat check to the script itself, regardless of the current working directory and without a dependency on README.md.